Phase 2.1: IBKR connection manager
- Add ib_insync dependency
- Implement IBKRClient with async connect/disconnect, heartbeat,
  and reconnection with backoff
- Add drain() method to EventBus for test synchronization
- Change reconnect_interval to float in BrokerConfig
- 27 tests covering connect, disconnect, reconnect, heartbeat,
  disconnect handler, and property access
- All quality gates pass (ruff, mypy, 104 tests)

+++ src/trading_platform/app/config.py
@@ -10,7 +10,7 @@ class BrokerConfig(BaseModel):
     port: int = 4001
     client_id: int = 1
     connect_timeout: int = 30
-    reconnect_interval: int = 5
+    reconnect_interval: float = 5.0
     max_reconnect_attempts: int = 10
